Detach thread on pthread cleanup
When the druntime runs the garbage collection it tries to suspend
all registered fails with leading to an onThreadError that crashes
from the Druntime when pthreads decide to cleanup a thread.  or
cleanup threads beforehand the thread_suspend command the program.
We have to cleanly detach the thread threads. However if the fuse
runtime suspends